backend/requirements.txt
Flask
gunicorn
google-cloud-datastore
flask_cors